+++ csvviz/csvviz.py
@@ -11,6 +11,7 @@ from typing import Dict as typeDict, List as typeList, Tuple as typeTuple, Union
 from typing import IO as typeIO
 
 
+from csvviz import __version__
 
 
 def clout(*args) -> typeNoReturn:
@@ -44,3 +45,15 @@ def clerr(*args) -> typeNoReturn:
 
 
 
+def print_version(ctx=None, param=None, value=None) -> typeNoReturn:
+    """
+    https://click.palletsprojects.com/en/3.x/options/#callbacks-and-eager-options
+    """
+    if not ctx:
+        clout(csvviz_version)
+    else:
+        # this is being used as a callback
+        if not value or ctx.resilient_parsing:
+            return
+        clout(csvviz_version)
+        ctx.exit()